refactor(cli): Move cli out of index file
